Rename FakeEvent to Event and move it to event.cc
