+++ SConscript
@@ -26,6 +26,7 @@ services_h = env.Command(['services.h'], ['services.py'], 'python3 ' + cereal_di
 
 messaging_objects = env.SharedObject([
   'messaging/messaging.cc',
+  'messaging/event.cc',
   'messaging/impl_zmq.cc',
   'messaging/impl_msgq.cc',
   'messaging/impl_fake.cc',
